Moving the Profile model to accounts/models.py
